# openapi/synapse.yaml
openapi: 3.1.0
info:
  title: Synapse API
  version: 0.1.0
  summary: Organizational Brain service for AI agents and team workflows.
  description: |
    Synapse is the governed layer that gives agents and teams permission-aware
    retrieval, a versioned skill registry, bounded memory, a user/team/org
    context service, a policy-guarded tool/connector (MCP) gateway, and durable
    workflow orchestration.

    Production identity is carried by a verified bearer JWT (JWKS, static RS256,
    or HS256). Trusted X-Principal-Id/X-Tenant-Id headers remain available for
    development or a trusted identity-aware gateway. Every governed endpoint is
    attributable for audit.

  license:
    name: Apache-2.0
    url: https://www.apache.org/licenses/LICENSE-2.0
  contact:
    name: ResQ Software
    url: https://github.com/resq-software/synapse
servers:
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /context.upsert:
    post:
      tags: [context]
      operationId: contextUpsert
      summary: Upsert a principal context profile.
      description: |
        Upserts a principal's context profile (personal/team/org). Personal
        memory is kept separate from org knowledge; PII is minimized and flagged
        via `data_classification`.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Context'
      responses:
        '200':
          description: Context upserted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ContextUpsertResponse'
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /audit/events:
    get:
      tags: [audit]
      operationId: auditEvents
      summary: Query the audit log.
      description: |
        Queries the append-only audit log, scoped to the caller's tenant. Every
        action that changes behavior or causes a side effect is recorded here.
      parameters:
        - $ref: '#/components/parameters/PrincipalId'
        - $ref: '#/components/parameters/TenantId'
        - $ref: '#/components/parameters/TeamIds'
        - $ref: '#/components/parameters/Role'
        - name: action
          in: query
          required: false
          description: Filter events by action name.
          schema:
            type: string
        - name: resource
          in: query
          required: false
          description: Filter events by resource identifier.
          schema:
            type: string
        - name: cursor
          in: query
          required: false
          description: Opaque pagination cursor returned by a previous response.
          schema:
            type: string
        - name: limit
          in: query
          required: false
          description: Maximum number of events to return.
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
      responses:
        '200':
          description: A page of audit events.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuditEventsResponse'
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Verified JWT; required when Synapse is configured for JWT authentication.
  parameters:
    PrincipalId:
      name: X-Principal-Id
      in: header
      required: false
      description: Trusted-header identity for development/gateway mode; ignored when JWT auth is configured.
      schema:
        type: string
      example: alice@resq
    TenantId:
      name: X-Tenant-Id
      in: header
      required: false
      description: Owning tenant/organization for this request.
      schema:
        type: string
      example: resq
    TeamIds:
      name: X-Team-Ids
      in: header
      required: false
      description: Comma-separated team ids the principal is acting within.
      schema:
        type: string
      example: team-ops,team-air
    Role:
      name: X-Role
      in: header
      required: false
      description: Platform/business role of the principal, used for RBAC.
      schema:
        type: string
      example: operator
  responses:
    BadRequest:
      description: The request was malformed or failed validation.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Missing or invalid bearer token or trusted principal.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Authenticated but not permitted by policy.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  schemas:
    Error:
      type: object
      description: Stable error envelope returned for all failure responses.
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable, machine-readable error code.
              examples: [not_found, bad_request, unauthorized, forbidden, db_error, internal]
            message:
              type: string
              description: Human-readable error message.
    DataClassification:
      type: object
      description: Data-classification flags that drive PII handling and retention.
      properties:
        contains_pii:
          type: boolean
          description: True if the context contains personally identifiable information.
        special_category:
          type: boolean
          description: True if the context contains special-category / sensitive personal data.
...
    Context:
      type: object
      description: A principal's user/team/org context profile (bounded, PII-minimized).
      required: [principal_id]
      properties:
        principal_id:
          type: string
          description: Stable identifier of the principal this context describes.
          example: user_9281
        tenant_id:
          type: string
          description: Owning tenant/organization.
          example: acme
        team_ids:
          type: array
          description: Team/group memberships.
          items:
            type: string
        role:
          type: string
          description: Platform/business role used for RBAC.
          example: manager
        location:
          type: string
          description: Coarse working location for routing / data-residency.
          example: US
        approval_limit_usd:
          type: number
          description: Max USD the principal may approve autonomously before human sign-off.
          example: 25000
        preferred_tools:
          type: array
          description: Tools/connectors the principal prefers.
          items:
            type: string
        active_projects:
          type: array
          description: Current workspace/project scopes.
          items:
            type: string
        policy_overrides:
          type: array
          description: Explicit policy overrides (named flags or structured rules).
          items: {}
        data_classification:
          $ref: '#/components/schemas/DataClassification'
        updated_at:
          type: string
          format: date-time
          description: RFC 3339 timestamp of the last update.
          example: '2026-07-01T00:00:00Z'
...
    AuditEvent:
      type: object
      description: An append-only audit event.
      required: [event_id, tenant_id, action, resource, outcome, ts]
      properties:
        event_id:
          type: string
          description: Stable event identifier.
        tenant_id:
          type: string
          description: Owning tenant/organization.
        principal_id:
          type: string
          description: Principal that triggered the event.
        action:
          type: string
          description: The action performed (e.g. tool.execute, skills.register).
        resource:
          type: string
          description: The resource acted upon.
        outcome:
          type: string
          description: Outcome of the action (e.g. allowed, denied, error).
        ts:
          type: string
          format: date-time
          description: RFC 3339 timestamp of the event.
          example: '2026-07-01T00:00:00Z'
        metadata:
          type: object
          description: Free-form event metadata (JSONB).
          additionalProperties: true
...
    ContextUpsertResponse:
      type: object
      description: Result of upserting a principal context.
      required: [principal_id, status]
      properties:
        principal_id:
          type: string
        status:
          type: string
          example: upserted
...
    AuditEventsResponse:
      type: object
      description: A page of audit events.
      required: [events]
      properties:
        events:
          type: array
          items:
            $ref: '#/components/schemas/AuditEvent'
        next_cursor:
          type: string
          description: Opaque cursor for the next page, when more results exist.
...